+++ install.sh
@@ -13,8 +13,7 @@ else
 fi
 
 # Files to install
-lib_files_and_dirs=("arctis_manager.py" "arctis_manager_launcher.py" "arctis_manager")
-bin_files=("bin/arctis-manager")
+bin_files=("dist/arctis-manager" "dist/arctis-manager-launcher")
 systemd_service_file="systemd/arctis-manager.service"
 udev_rules_file="udev/91-steelseries-arctis.rules"
 desktop_file="ArctisManager.desktop"
@@ -24,8 +23,6 @@ icon_file="arctis_manager/images/steelseries_logo.svg"
 applications_dir="${chroot_path}${install_prefix}/share/applications/"
 icons_dir="${chroot_path}/usr/share/icons/hicolor/scalable/apps/"
 bin_dir="${chroot_path}${install_prefix}/bin"
-lib_dir="${install_prefix}/lib/arctis-manager"
-chrooted_lib_dir="${chroot_path}${lib_dir}"
 udev_dir="${chroot_path}/usr/lib/udev/rules.d/"
 systemd_dir="${chroot_path}/usr/lib/systemd/user/"
 
@@ -41,58 +38,29 @@ function install() {
     echo "Installing Arctis Manager..."
 
     superuserdo mkdir -p "${bin_dir}"
-    superuserdo mkdir -p "${chrooted_lib_dir}"
     superuserdo mkdir -p "${applications_dir}"
     superuserdo mkdir -p "${icons_dir}"
+
+    echo "Running pyinstaller to generate binary files"
+    python3 -m pip install --upgrade pipenv
+    python -m pipenv install -d
+    python -m pipenv run pyinstaller arctis-manager.spec
+    python -m pipenv run pyinstaller arctis-manager-launcher.spec
+    python -m pipenv --rm
 
     echo "Installing binaries in ${bin_dir}"
     for file in "${bin_files[@]}"; do
         dest_file="${bin_dir}/$(basename "${file}")"
         superuserdo cp "${file}" "${dest_file}"
-
-        # Replace placeholders
-        superuserdo sed -i "s|{{LIBDIR}}|${lib_dir}|g" "${dest_file}"
     done
 
     echo "Installing desktop file in ${applications_dir}"
     dest_file="${applications_dir}/$(basename "${desktop_file}")"
     superuserdo cp "${desktop_file}" "${dest_file}"
-    # Replace placeholders
-    superuserdo sed -i "s|{{LIBDIR}}|${lib_dir}|g" "${dest_file}"
 
     echo "Installing icon file in ${icons_dir}"
     dest_file="${icons_dir}/arctis_manager.svg"
     superuserdo cp "${icon_file}" "${dest_file}"
-
-    echo "Installing application data in ${chrooted_lib_dir}"
-    for file in "${lib_files_and_dirs[@]}"; do
-        if [ -f "${file}" ]; then
-            superuserdo cp "${file}" "${chrooted_lib_dir}/"
-        elif [ -d "${file}" ]; then
-            superuserdo cp -r "${file}" "${chrooted_lib_dir}/"
-        fi
-    done
-
-    # Note: using superuserdo with pip to install modules
-    # It is harmless, as it will target a custom folder
-    echo "Installing python dependencies in ${chrooted_lib_dir}"
-    echo "pip configuration:"
-    if [ -n "${chroot_path}" ]; then
-        printf "\troot: %s\n" "${chroot_path}"
-        printf "\tprefix: %s\n" "${lib_dir}"
-        superuserdo python3 -m pip install \
-            --upgrade --quiet --root-user-action=ignore \
-            --config-settings="--install-scripts=${bin_dir}" \
-            -r requirements.txt \
-            --no-binary 'pylupdate6' \
-            --root "${chroot_path}" --prefix "${lib_dir}"
-        
-    else
-        printf "\ttarget: %s\n" "${lib_dir}"
-        superuserdo python3 -m pip install --upgrade --quiet --root-user-action=ignore -r requirements.txt --target "${lib_dir}"
-    fi
-
-    superuserdo find "${chroot_path}${lib_dir}" -name "__pycache__" -exec rm -rf {} \;
 
     # Udev rules
     echo "Installing udev rules."